pyproject.toml
[project]
name = "ghost-crew"
version = "0.1.0"
description = "Your invisible AI crew that handles Slack requests on your behalf"
requires-python = ">=3.11"
dependencies = [
    "slack-bolt>=1.18.0",
    "slack-sdk>=3.27.0",
    "anthropic>=0.40.0",
    "google-api-python-client>=2.100.0",
    "google-auth-oauthlib>=1.1.0",
    "pyyaml>=6.0",
    "chromadb>=0.4.0",
    "gitpython>=3.1.0",
    "apscheduler>=3.10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.backends._legacy:_Backend"
